requirements.txt
# Web scraping
selenium>=4.15.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Configuration and utilities
PyYAML>=6.0.0
loguru>=0.7.0

# Date and time handling
python-dateutil>=2.8.0

# Optional: For advanced text processing
beautifulsoup4>=4.12.0
lxml>=4.9.0